pyproject.toml
[project]
name = "irs-telegram-bot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "groq==0.9.0",
    "openai==1.3.0",
    "python-dotenv==1.0.0",
    "python-telegram-bot==20.7",
    "requests>=2.32.5",
]
